requirements.txt
spotdl
pyTelegramBotAPI
python-dotenv
requests